Se limpio la db, se agrego un exp_date a Coupon, tracking_code a Sale y se modificaron los campos de date en Planning que ahora es format_date, amount en RecipeDetail que ahora es Decimal

+++ foodyplus/foodyplus/foods/models/recipe_detail.py
@@ -15,10 +15,12 @@ class RecipeDetail(BasicModel):
 
     product = models.ForeignKey('Product', on_delete=models.CASCADE)
 
-    amount = models.PositiveIntegerField(
+    amount = models.DecimalField(
         "Cantidad del producto",
         default=1,
-        blank=True
+        blank=True,
+        max_digits=19,
+        decimal_places=2,
     )
 
     unit = models.CharField(
